// scripts/src/Enums/SkillsSurveyPageName.ts
/**
 * Skills Survey Page Name - Each value represents the name of a Skills Survey Page as defined in the SurveyJS
 * JSON specification.
 */
export enum SkillsSurveyPageName {
    // All new entries should be added in alphabetical order
    AdditionalInfo = "Page.AdditionalInfo",
    Certification = "Page.Certification",
    DIJurisdiction = "Page.DIJurisdiction",
    Education = "Page.Education",
    IndustryExperience = "Page.IndustryExperience",
    IndustryManufacturing = "Page.IndustryManufacturingExperience",
    InterimMgmt = "Page.InterimMgmt",
    Language = "Page.Language",
    Manufacturing = "Page.ManufacturingExperience",
    NACR = "Page.NACR",
    SecurityClearance = "Page.SecurityClearance",
    SoftwareTool = "Page.SoftwareTool",
    SolutionIndustry = "Page.SolutionIndustry",
    SolutionSkills = "Page.SolutionSkills",
    Sourcing = "Page.SourcingCategory",
    WorkExperience = "Page.WorkExperience",
}